// src/api/generated/models/templateFieldResponse.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * Gaula Digital — API
 * Sistema de información operacional del GAULA, Ejército Nacional de Colombia.
 * OpenAPI spec version: v1
 */
import type { TemplateFieldResponseType } from './templateFieldResponseType';

export interface TemplateFieldResponse {
  code?: string;
  label?: string;
  type?: TemplateFieldResponseType;
  required?: boolean;
  min?: number;
  max?: number;
  minLength?: number;
  maxLength?: number;
  pattern?: string;
  options?: string[];
  notFuture?: boolean;
  fields?: TemplateFieldResponse[];
}